Etiquetas service feito, comentários adicionados

+++ src/main/java/com/example/test/demo/service/EtiquetaService.java
@@ -1,8 +1,6 @@
 package com.example.test.demo.service;
 
-import com.example.test.demo.model.Componente;
 import com.example.test.demo.model.EtiquetaMaterial;
-import com.example.test.demo.model.Material;
 import com.google.api.core.ApiFuture;
 import com.google.cloud.firestore.Firestore;
 import com.google.cloud.firestore.QueryDocumentSnapshot;
@@ -19,6 +17,8 @@ import java.util.concurrent.ExecutionException;
 public class EtiquetaService {
 
     public static final String COL_NAME = "etiquetaMaterial";
+    public static final String COL_NAME_COMPONENT = "component";
+    public static final String COL_NAME_MATERIAL = "material";
 
     /**
      * Metodo cria uma nova etiqueta na base de dados
@@ -27,7 +27,10 @@ public class EtiquetaService {
      * @throws ExecutionException
      * @throws InterruptedException
      */
-    public String saveEtiqueta(EtiquetaMaterial etiquetaMaterial) throws ExecutionException, InterruptedException {
+    public String createEtiqueta(EtiquetaMaterial etiquetaMaterial) throws ExecutionException, InterruptedException {
+        if (verifyEtiqueta(etiquetaMaterial)) {
+            return null;
+        }
         Firestore db = FirestoreClient.getFirestore();
         ApiFuture<QuerySnapshot> future = db.collection(COL_NAME).get();
         List<QueryDocumentSnapshot> documents = future.get().getDocuments();
@@ -39,18 +42,18 @@ public class EtiquetaService {
             oldEtiqueta = doc.toObject(EtiquetaMaterial.class);
             if (oldEtiqueta.getEtiquetaId() > biggest) {
                 biggest = oldEtiqueta.getEtiquetaId();
-
             }
         }
         etiquetaMaterial.setEtiquetaId(biggest + 1);
         /*ADICIONA UM NOVO MATERIAL*/
         ApiFuture<WriteResult> colApiFuture = db.collection(COL_NAME).document().set(etiquetaMaterial);
-
         return "etiqueta created";
     }
 
+
     /**
      * Metodo que retorna uma lista de etiquetas
+     *
      * @return retorna uma lista de etiquetas
      * @throws ExecutionException
      * @throws InterruptedException
@@ -60,36 +63,40 @@ public class EtiquetaService {
         ApiFuture<QuerySnapshot> future = db.collection(COL_NAME).get();
         List<QueryDocumentSnapshot> documents = future.get().getDocuments();
         List<EtiquetaMaterial> mats = new ArrayList<>();
-        if (documents.size() > 0) {
-            for (QueryDocumentSnapshot doc : documents) {
-                mats.add(doc.toObject(EtiquetaMaterial.class));
-            }
-            return mats;
+        if (documents.isEmpty())
+            return null;
+        for (QueryDocumentSnapshot doc : documents) {
+            mats.add(doc.toObject(EtiquetaMaterial.class));
         }
-        return null;
+        return mats;
     }
 
 
     /**
-     * metodo que atualiza uma etiqueta
+     * metodo que atualiza uma etiqueta com um id especifico
+     *
      * @param etiquetaMaterial objeto enviado do client side
      * @return retorna a data de atualização da etiqueta
      * @throws ExecutionException
      * @throws InterruptedException
      */
     public String updateEtiqueta(EtiquetaMaterial etiquetaMaterial) throws ExecutionException, InterruptedException {
+        if (verifyEtiqueta(etiquetaMaterial)) {
+            return null;
+        }
         Firestore db = FirestoreClient.getFirestore();
         ApiFuture<QuerySnapshot> future = db.collection(COL_NAME).whereEqualTo("etiquetaId", etiquetaMaterial.getEtiquetaId()).get();
         //update a document from firestore
-        if (future.get().size() <= 0)
-            return "No elements to be queried";
-        ApiFuture<WriteResult> apiFuture = db.collection(COL_NAME).document(future.get().getDocuments().get(0).getId()).set(etiquetaMaterial);
 
-        return "updated";
+        if (future.get().size() <= 0)
+            return null;
+        ApiFuture<WriteResult> apiFuture = db.collection(COL_NAME).document(future.get().getDocuments().get(0).getId()).set(etiquetaMaterial);
+        return "etiqueta updated:" + etiquetaMaterial.getEtiquetaId();
     }
 
     /**
-     * metodo que apaga uma etiqueta
+     * metodo que apaga uma etiqueta na base de dados
+     *
      * @param id id da etiqueta a ser apagada
      * @return retorna a data de apagamento da etiquetas
      * @throws ExecutionException
@@ -97,29 +104,32 @@ public class EtiquetaService {
      */
 
     public String deleteEtiqueta(int id) throws ExecutionException, InterruptedException {
+        if (id < 0)
+            return null;
         Firestore db = FirestoreClient.getFirestore();
         ApiFuture<QuerySnapshot> future = db.collection(COL_NAME).whereEqualTo("etiquetaId", id).get();
         if (future.get().size() <= 0)
-            return "Material não encontrado para ser eleminado";
-        System.out.println(db.collection(COL_NAME).document(future.get().getDocuments().get(0).getId()));
+            return null;
         ApiFuture<WriteResult> writeResult = db.collection(COL_NAME).document(future.get().getDocuments().get(0).getId()).delete();
-        return writeResult.get().getUpdateTime().toString();
+        return "etiqueta deleted with:" + id;
     }
+
     /*CASOS PARTICULARES*/
     /*
      * COMPONENTES:*ADICIONAR COMPONENTE À ETIQUETA
-     *APAGAR COMPONENTE À ETIQUETA
-     *ALTERAR COMPONENTE À ETIQUETA
+     * APAGAR COMPONENTE À ETIQUETA
+     * ALTERAR COMPONENTE À ETIQUETA
      * LER TODOS OS COMPONENTES DA ETIQUETA
      * MATERIAIS:*ADICIONAR MATERIAL À ETIQUETA
-     *LER MATERIAIS DA ETIQUETA
+     * LER MATERIAIS DA ETIQUETA
      * APAGAR MATERIAL DA ETIQUETA
      */
 
     /**
      * metodo que adiciona um componente a uma etiqueta
+     *
      * @param idEtiqueta id da etiqueta
-     * @param comp componente a ser adicionado
+     * @param comp       componente a ser adicionado
      * @return retorna a data de atualização da etiqueta
      * @throws ExecutionException
      * @throws InterruptedException
@@ -139,6 +149,7 @@ public class EtiquetaService {
 
     /**
      * metodo apaga um componente de uma etiqueta
+     *
      * @param idEtiqueta id da etiqueta
      * @param componente componente a ser apagado
      * @return retorna a data de atualização da etiqueta
@@ -165,6 +176,7 @@ public class EtiquetaService {
 
     /**
      * metodo que devolve todos os componentes de uma etiqueta
+     *
      * @param idEtiqueta id da etiqueta
      * @return retorna uma lista de componentes
      * @throws ExecutionException
@@ -183,6 +195,7 @@ public class EtiquetaService {
 
     /**
      * metodo que adiciona um material a uma etiqueta
+     *
      * @param idEtiqueta id da etiqueta
      * @param idMaterial id do material a ser adicionado
      * @return retorna a data de atualização da etiqueta
@@ -204,6 +217,7 @@ public class EtiquetaService {
 
     /**
      * metodo que apaga um material de uma etiqueta
+     *
      * @param idEtiqueta id da etiqueta
      * @param idMaterial id do material a ser apagado
      * @return retorna a data de atualização da etiqueta
@@ -225,6 +239,7 @@ public class EtiquetaService {
 
     /**
      * metodo que devolve todos os materiais de uma etiqueta
+     *
      * @param idEtiqueta
      * @return
      * @throws ExecutionException
@@ -238,6 +253,75 @@ public class EtiquetaService {
             return null;
         EtiquetaMaterial etiquetaMaterial = future.get().getDocuments().get(0).toObject(EtiquetaMaterial.class);
         return etiquetaMaterial.getMateriaisId();
+    }
+    /*METODOS AUXILIARES*/
+
+    /**
+     * metodo verifica se as componentes da lista são validas
+     * @param idComponente id da componente a ser verificada
+     * @return retorna verdadeiro se for invalida
+     * @throws ExecutionException
+     * @throws InterruptedException
+     */
+    private boolean verifyComponentes(Integer idComponente) throws ExecutionException, InterruptedException {
+        Firestore db = FirestoreClient.getFirestore();
+        ApiFuture<QuerySnapshot> future = db.collection(COL_NAME_COMPONENT).whereEqualTo("id", idComponente).get();
+        if (future.get().isEmpty())
+            return true;
+        return false;
+    }
+
+    /**
+     * metodo que verifica se os materiais da lista são validos
+     * @param material id do material a ser verificado
+     * @return retorna verdadeiro se for invalido
+     * @throws ExecutionException
+     * @throws InterruptedException
+     */
+    private boolean verifyMateriais(Integer material) throws ExecutionException, InterruptedException {
+        Firestore db = FirestoreClient.getFirestore();
+        ApiFuture<QuerySnapshot> future = db.collection(COL_NAME_MATERIAL).whereEqualTo("id", material).get();
+        if (future.get().isEmpty())
+            return true;
+        return false;
+    }
+
+    /**
+     * metodo que verifica se a etiqueta é valida
+     * @param etiquetaMaterial objeto a ser verificado
+     * @return retorna verdadeiro se for invalida
+     * @throws ExecutionException
+     * @throws InterruptedException
+     */
+    private boolean verifyEtiqueta(EtiquetaMaterial etiquetaMaterial) throws ExecutionException, InterruptedException {
+        return (etiquetaMaterial.getDescricaoMaterial() == null || etiquetaMaterial.getDescricaoMaterial().length() > 128 || etiquetaMaterial.getDescricaoMaterial().length() < 1)
+                || (etiquetaMaterial.getQuantidade() < 0 || etiquetaMaterial.getQuantidade() > 99)
+                || (etiquetaMaterial.getSubEtiqueta() == null || etiquetaMaterial.getSubEtiqueta().length() > 32 || etiquetaMaterial.getSubEtiqueta().length() < 1)
+                || (etiquetaMaterial.getEtiquetaId() < 0)
+                || (etiquetaMaterial.getEtiqueta() == null || etiquetaMaterial.getEtiqueta().length() > 32 || etiquetaMaterial.getEtiqueta().length() < 1)
+                || (etiquetaMaterial.getComponentes() == null || etiquetaMaterial.getMateriaisId() == null || verifyArrays(etiquetaMaterial))
+                || (!etiquetaMaterial.getEtiqueta().equalsIgnoreCase("NÃO CONSUMIVEIS") && !etiquetaMaterial.getEtiqueta().equalsIgnoreCase("CONSUMIVEIS"));
+    }
+
+    /**
+     * metodo que verifica se os arraysLists são validos
+     * @param etiquetaMaterial objeto a ser verificado
+     * @return retorna verdadeiro se pelo menos 1 for invalido
+     * @throws ExecutionException
+     * @throws InterruptedException
+     */
+    private boolean verifyArrays(EtiquetaMaterial etiquetaMaterial) throws ExecutionException, InterruptedException {
+        for (Integer componente : etiquetaMaterial.getComponentes()) {
+            if (componente == null || componente < 0 || verifyComponentes(componente)) {
+                return true;
+            }
+        }
+        for (Integer material : etiquetaMaterial.getMateriaisId()) {
+            if (material == null || material < 0 || verifyMateriais(material)) {
+                return true;
+            }
+        }
+        return false;
     }
 
 }
